Move tests back to integration_tests.
